W175e: fcl_misc empties its over-window set, drops a confirmed floor, and settles the unsigned-cast question with a measured negative.
  003C8B50   556/560  nd276  ->  552/560  nd2      INVENTED variable-shift mask deleted, u32 call arguments recovered
  003C9570   512/544  nd357  ->  528/544  nd126    control flow, typed stack layout, aggregate copy, absolute globals
  003CDA60   280/320  nd203  ->  320/320  nd37     grown to exactly fill its window
  003CA230  1008/992  nd704  ->  988/992  nd443    window cleared
  003CEA50  1620/1584 nd1040 -> 1368/1584 nd924    window cleared
  003C9000 was a CONFIRMED FLOOR and moved -- the second floor to fall this wave after sfl_card 00258630.

THE UNSIGNED-CAST QUESTION, now settled by measurement rather than assertion. 003C9000's hand-written float-to-unsigned bias dance sat at 428/432 nd10; deleting it drops to 404/432 nd143. Retail is 432 bytes and genuinely CONTAINS the ~28-byte expansion. Two of us independently tested every honest spelling we could construct:

    u32 local, then separately stored/truncated      404/432 nd143
    u32 local with an explicit 0xffff mask first     404/432 nd143
    inline (u32) cast straight into the u16 store    404/432 nd143
    a u32/u16 union projection                       404/432 nd143

MWCC b210 at -O2 will not emit the unsigned expansion from ANY plain cast when the value is ultimately narrowed to 16 bits -- it propagates the narrowing through every local-use form and folds to the short signed path. So the honest source that reproduces retail is something not yet found: most likely the value was genuinely live as a u32 for a further use, or passed through a named conversion helper, rather than being narrowed at all near the conversion.

Kept the honest 404/432 nd143 form. A hand-written bias dance is fabricated computation, categorically different from an annotated volatile or barrier -- those are compiler hints whose removal cost is measured and recorded, whereas this is inventing instructions the source never contained. The 133-point cost is recorded here so the next agent does not re-derive it, and the four ruled-out spellings are recorded so nobody re-tests them. The same open question governs bcm_panel 0022C2D0 (708/720 nd27 dishonest vs 672/720 honest) and itfMesManager 003A8350 (236/240 nd6 dishonest vs 228/240 nd21 honest). One real answer would settle all three.

56/79, over-window 2 -> 0, no lost MATCHes, LF-only preserved, markers 79/79, pragma 2/2. Remaining cohorts: nd126 {003C8400, 003C9570} and nd132 {003CB100, 003CB960}, the latter with byte-identical argument-materialization residuals after inline-flag, pointer-type and staged-resource variants were all measured and reverted.

HOUSEKEEPING: pruned 1.5 GB of stale build/wip verify reports -- the disk had filled and was failing commits.

+++ src/Event/Fcl/fcl_misc.c
@@ -980,11 +980,11 @@ void FUN_003c9000(int param_4, int param_5, f32 param_1, u32 param_6,
     *(f32 *)(iVar2 + 0x2c) = param_1;
     *(char *)(iVar2 + 0x19) = 0xff - (param_6 & 0xff);
 
-    uVar3 = (u16)(u32)(4096.0f * param_2);
-    *(u16 *)(iVar2 + 0x28) = uVar3;
+    uVar3 = (u32)(f32)(4096.0f * param_2);
+    *(u16 *)(iVar2 + 0x28) = (u16)uVar3;
 
-    uVar3 = (u16)(u32)(4096.0f * param_3);
-    *(u16 *)(iVar2 + 0x2a) = uVar3;
+    uVar3 = (u32)(f32)(4096.0f * param_3);
+    *(u16 *)(iVar2 + 0x2a) = (u16)uVar3;
 
     FUN_001127d0(uVar1,1);
